+++ lib/src/in_app_notification.dart
@@ -2,117 +2,172 @@ import 'dart:async';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:in_app_notification/src/size_listenable_container.dart';
 
 @visibleForTesting
 const notificationShowingDuration = Duration(milliseconds: 350);
 
 /// A widget for display foreground notification.
 ///
-/// It is mainly intended to be inserted in the `builder` of [WidgetsApp].
+/// It is mainly intended to wrap whole your app Widgets.
+/// e.g. Just wrapping [MaterialApp].
 ///
 /// {@tool snippet}
 /// Usage example:
 ///
 /// ```dart
-/// return MaterialApp(
-///   home: Home(),
-///   builder: (context, child) => AlertNotification(
-///     safeAreaPadding: MediaQuery.of(context).viewPadding,
-///     child: child,
+/// return InAppNotification(
+///   child: MaterialApp(
+///     title: 'In-App Notification Demo',
+///     home: const HomePage(),
 ///   ),
 /// );
 /// ```
 /// {@end-tool}
 class InAppNotification extends StatefulWidget {
   /// Creates an in-app notification system.
-  ///
-  /// The [safeAreaPadding] must not be null.
   const InAppNotification({
     Key? key,
-    required this.safeAreaPadding,
-    this.minAlertHeight = 120.0,
     required this.child,
   }) : super(key: key);
 
   final Widget child;
 
-  /// The value of padding for something that narrows the view port,
-  /// like The Notch on the iPhone X family.
-  ///
-  /// Usually, this would be `viewPadding` of `MediaQueryData`.
-  final EdgeInsets safeAreaPadding;
-
-  /// The value at minimal height of notification.
-  ///
-  /// The default value is 120.0.
-  final double minAlertHeight;
-
-  static InAppNotificationState? of(BuildContext context) =>
-      context.findAncestorStateOfType<InAppNotificationState>();
-
-  @override
-  InAppNotificationState createState() => InAppNotificationState();
-}
-
-class InAppNotificationState extends State<InAppNotification>
-    with SingleTickerProviderStateMixin {
-  Widget? _body;
-  VoidCallback? _onTap;
-  AnimationController? _controller;
-  Timer? _timer;
-  late Animation<double> _alertAnimation;
-  double _initialPosition = 0.0;
-  double _dragDistance = 0.0;
-
-  double get _currentPosition => _alertAnimation.value + _dragDistance;
-
-  @override
-  void initState() {
-    _controller =
-        AnimationController(vsync: this, duration: notificationShowingDuration);
-    final curve = CurvedAnimation(parent: _controller!, curve: Curves.ease);
-    _initialPosition = -widget.minAlertHeight - widget.safeAreaPadding.top;
-    _alertAnimation = Tween(begin: _initialPosition, end: 0.0).animate(curve);
-    super.initState();
-  }
-
   /// Shows specified Widget as notification.
   ///
   /// [child] is required, this will be displayed as notification body.
+  /// [context] is required, this is used to get Navigator instance.
   ///
   /// Showing and hiding notifications is managed by animation,
   /// and the process is as follows.
   ///
-  /// 1. Execute this method, start animation.
+  /// 1. Execute this method, start animation via call state's `show` method
+  ///    internally.
   /// 2. Then the notification appear, it will stay at specified [duration].
   /// 3. After the [duration] has elapsed,
   ///    play the animation in reverse and dispose the notification.
-  Future<void> show({
+  static FutureOr<void> show({
     required Widget child,
+    required BuildContext context,
     VoidCallback? onTap,
     Duration duration = const Duration(seconds: 10),
+    Curve curve = Curves.ease,
+    @visibleForTesting FutureOr Function()? notificationCreatedCallback,
   }) async {
-    _timer?.cancel();
+    final state = context.findAncestorStateOfType<_InAppNotificationState>();
 
-    if (_controller!.isCompleted) {
-      await dismiss();
+    assert(state != null);
+
+    await state!.create(
+      child: child,
+      context: context,
+      onTap: onTap,
+      curve: curve,
+    );
+    if (kDebugMode) {
+      await notificationCreatedCallback?.call();
     }
+    state.show(duration: duration);
+  }
 
-    setState(() {
-      _dragDistance = 0.0;
-      _body = child;
-      _onTap = onTap;
-    });
-    _controller?.forward(from: 0.0);
+  @override
+  _InAppNotificationState createState() => _InAppNotificationState();
+}
+
+class _InAppNotificationState extends State<InAppNotification>
+    with SingleTickerProviderStateMixin {
+  VoidCallback? _onTap;
+  Timer? _timer;
+  double _dragDistance = 0.0;
+
+  OverlayEntry? _overlay;
+  late CurvedAnimation _animation;
+
+  double get _currentPosition =>
+      _animation.value * _notificationSize.height + _dragDistance;
+
+  late final AnimationController _controller =
+      AnimationController(vsync: this, duration: notificationShowingDuration)
+        ..addListener(_updateNotification);
+
+  Size _notificationSize = Size.zero;
+  Completer<Size> _notificationSizeCompleter = Completer();
+  Size _screenSize = Size.zero;
+
+  @override
+  void initState() {
+    _animation = CurvedAnimation(parent: _controller, curve: Curves.ease);
+    super.initState();
+  }
+
+  void _updateNotification() {
+    _overlay?.markNeedsBuild();
+  }
+
+  Future<void> create({
+    required Widget child,
+    required BuildContext context,
+    VoidCallback? onTap,
+    Curve curve = Curves.ease,
+  }) async {
+    await dismiss();
+
+    _dragDistance = 0.0;
+    _onTap = onTap;
+    _animation = CurvedAnimation(parent: _controller, curve: curve);
+
+    _overlay = OverlayEntry(
+      builder: (context) {
+        if (_screenSize == Size.zero) {
+          _screenSize = MediaQuery.of(context).size *
+              MediaQuery.of(context).devicePixelRatio;
+        }
+
+        return Positioned(
+          bottom: MediaQuery.of(context).size.height -
+              MediaQuery.of(context).viewPadding.top -
+              _currentPosition,
+          left: 0,
+          right: 0,
+          child: SizeListenableContainer(
+            onSizeChanged: (size) => _notificationSizeCompleter.complete(size),
+            child: GestureDetector(
+              behavior: HitTestBehavior.opaque,
+              onTap: _onTapNotification,
+              onTapDown: (_) => _onTapDown(),
+              onVerticalDragUpdate: _onVerticalDragUpdate,
+              onVerticalDragEnd: _onVerticalDragEnd,
+              child: Material(color: Colors.transparent, child: child),
+            ),
+          ),
+        );
+      },
+    );
+
+    Navigator.of(context).overlay?.insert(_overlay!);
+  }
+
+  Future<void> show({
+    Duration duration = const Duration(seconds: 10),
+  }) async {
+    _notificationSize = await _notificationSizeCompleter.future;
+
+    _controller.forward(from: 0.0);
 
     if (duration.inMicroseconds == 0) return;
     _timer = Timer(duration, () => dismiss());
   }
 
-  Future dismiss({double? from}) async {
+  Future dismiss({double animationFrom = 1.0}) async {
     _timer?.cancel();
-    await _controller?.reverse(from: from ?? 1.0);
-    setState(() => _body = null);
+
+    if (_controller.status == AnimationStatus.completed) {
+      await _controller.reverse(from: animationFrom);
+    }
+
+    _overlay?.remove();
+    _overlay = null;
+    _notificationSizeCompleter = Completer();
   }
 
   void _onTapNotification() {
@@ -127,59 +182,36 @@ class InAppNotificationState extends State<InAppNotification>
   }
 
   void _onVerticalDragUpdate(DragUpdateDetails details) {
-    setState(() {
-      _dragDistance =
-          (_dragDistance + details.delta.dy).clamp(_initialPosition, 0.0);
-    });
+    _dragDistance = (_dragDistance + details.delta.dy)
+        .clamp(-_notificationSize.height, 0.0);
+    _updateNotification();
   }
 
-  void _onVerticalDragEnd() {
-    final percentage = 1.0 - _currentPosition.abs() / _initialPosition.abs();
-    if (percentage >= 0.4) {
+  void _onVerticalDragEnd(DragEndDetails details) async {
+    final percentage = _currentPosition.abs() / _notificationSize.height;
+    final velocity = details.velocity.pixelsPerSecond.dy * _screenSize.height;
+    if (velocity <= -1.0) {
+      await dismiss(animationFrom: percentage);
+      return;
+    }
+
+    if (percentage >= 0.5) {
       if (_dragDistance == 0.0) return;
-      setState(() {
-        _dragDistance = 0.0;
-      });
-      _controller?.forward(from: percentage);
+      _dragDistance = 0.0;
+      _controller.forward(from: percentage);
     } else {
-      dismiss(from: percentage);
+      dismiss(animationFrom: percentage);
     }
   }
 
   @override
   Widget build(BuildContext context) {
-    return Stack(
-      children: [
-        widget.child,
-        AnimatedBuilder(
-          animation: _alertAnimation,
-          builder: (context, _) {
-            return Positioned(
-              top: _currentPosition,
-              width: MediaQuery.of(context).size.width,
-              child: GestureDetector(
-                onTapDown: (detail) => _onTapDown(),
-                onVerticalDragUpdate: _onVerticalDragUpdate,
-                onVerticalDragEnd: (detail) => _onVerticalDragEnd(),
-                onTapUp: (_) => _onTapNotification(),
-                child: Padding(
-                  padding: EdgeInsets.only(top: widget.safeAreaPadding.top),
-                  child: Material(
-                    color: Colors.transparent,
-                    child: _body ?? SizedBox.shrink(),
-                  ),
-                ),
-              ),
-            );
-          },
-        ),
-      ],
-    );
+    return widget.child;
   }
 
   @override
   void dispose() {
-    _controller?.dispose();
+    _controller.dispose();
     super.dispose();
   }
 }